refactor(web): split Group 1 routes (keys/connectors/pages/activity/notifications/push/backup) into APIRouters
ev/interfaces/web/app.py's create_app() held all ~177 routes as closures in
one function. Introduces a WebContext (config/memory/brain/commands/owner +
the check/body/recurval/folders/conv/status_text/run_command/env_write
helpers, moved verbatim from create_app()'s closures) and extracts the
lowest-risk, lowest-traffic domain routes into ev/interfaces/web/routes/
APIRouters that take a WebContext instead of closing over local variables.
Group 1 of 5 for this phase — CRUD principal, session/auth/oauth, external
APIs, and aggregation/rest routes are unchanged, staying inline in app.py for
future PRs.

No route path, method, or response shape changed: verified by diffing
app.openapi()'s full path/method table (main vs this branch — 177/177
identical) and probing every touched endpoint's status/response shape with
matching TestClient instances of both versions.
